add MultisigEndedMessageFactory
