+++ novawallet/Common/PushHandling/Multisig/MultisigEndedMessageModel.swift
